refactor: add folder for default data
